Moved `Quest.delete` to `QuestDB.deleteQuest`
