+++ src/view/log/HandlerLog.js
@@ -28,8 +28,7 @@ export default class HandlerLog
       const result = await FQLDialog.confirmDeleteQuest({ name, result: questId, questId, isQuestLog: true });
       if (result)
       {
-         const deleteQuest = QuestDB.getQuest(result);
-         if (deleteQuest) { await Socket.deletedQuest(await deleteQuest.delete()); }
+         await QuestDB.deleteQuest({ questId: result });
       }
    }
